# apps/api/openapi/bezzo-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/buyer/addresses/{addressId}": {
      "get": {
        "operationId": "BuyersController_address",
        "parameters": [
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "buyer"
        ]
      },
      "patch": {
        "operationId": "BuyersController_updateAddress",
        "parameters": [
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a delivery address",
        "tags": [
          "buyer"
        ]
      },
      "delete": {
        "operationId": "BuyersController_deleteAddress",
        "parameters": [
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Archive a delivery address",
        "tags": [
          "buyer"
        ]
      }
    },
...
    "/api/v1/buyer/documents/{documentId}": {
      "delete": {
        "operationId": "BuyersController_deleteDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Remove a pending or rejected document",
        "tags": [
          "buyer"
        ]
      }
    },
...
    "/api/v1/supplier/listings/{listingId}": {
      "get": {
        "operationId": "SuppliersController_listing",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get one listing with stock and pricing",
        "tags": [
          "supplier"
        ]
      },
      "patch": {
        "operationId": "SuppliersController_updateListing",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update price, MOQ, lead time or listing status (price changes are versioned)",
        "tags": [
          "supplier"
        ]
      }
    },
...
    "/api/v1/supplier/inventory/{inventoryId}/adjust": {
      "post": {
        "operationId": "SuppliersController_adjustStock",
        "parameters": [
          {
            "name": "inventoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Adjust available stock by a delta; every change is written to the stock ledger",
        "tags": [
          "supplier"
        ]
      }
    },
    "/api/v1/supplier/inventory/{inventoryId}/set": {
      "post": {
        "operationId": "SuppliersController_setStock",
        "parameters": [
          {
            "name": "inventoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Set available stock to an absolute value (stock-take); reserved stock is preserved",
        "tags": [
          "supplier"
        ]
      }
    },
    "/api/v1/supplier/inventory/{inventoryId}/ledger": {
      "get": {
        "operationId": "SuppliersController_ledger",
        "parameters": [
          {
            "name": "inventoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Append-only stock ledger for one inventory position",
        "tags": [
          "supplier"
        ]
      }
    },
...
    "/api/v1/catalog/products/{productId}": {
      "get": {
        "operationId": "CatalogController_product",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Product detail with live supplier offers (best price first)",
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "BEZZO API",
    "description": "B2B pharmaceutical marketplace API. Standardised envelopes, stable error codes, Idempotency-Key on mutations, role + permission + ownership authorization.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}